code commit
A common problem that we face in RESTful web services is sometimes the search API returns huge data.
E.g. Searching for users in a city may return thousands of users. Returning this while result data consumes lot of bandwidth and
not best way. We will see how to use PagingAndSortingRepository of Spring Data to address this issue.